Refactoring, moved cql library and measure creation to it own file. And some more small fixes
