+++ cypress/integration/MAT/WebInterface/MeasureComposer/MeasurePackager/ErrorMessageMeasurePackaging.spec.js
@@ -2,6 +2,7 @@ import * as helper from "../../../../../support/helpers";
 import * as measurelibrary from "../../../../../pom/MAT/WI/MeasureLibrary";
 import * as measureComposer from "../../../../../pom/MAT/WI/MeasureComposer";
 import * as oktaLogin from "../../../../../support/oktaLogin";
+import * as dataCreation from "../../../../../support/MAT/MeasureAndCQLLibraryCreation";
 
 let fhirMeasure = ''
 let qdmMeasure = ''
@@ -10,8 +11,10 @@ describe('Measure Packager: Validate before packaging a FHIR measure', () => {
     before('Login', () => {
         oktaLogin.login()
 
-        qdmMeasure = helper.createDraftMeasure('qdmDraftMeasure', 'QDM')
-        fhirMeasure = helper.createDraftMeasure('FhirDraftMeasure', 'FHIR')
+        qdmMeasure = dataCreation.createDraftMeasure('qdmDraftMeasure', 'QDM')
+        fhirMeasure = dataCreation.createDraftMeasure('FhirDraftMeasure', 'FHIR')
+
+        helper.verifySpinnerAppearsAndDissappears()
 
     })
     beforeEach('Preserve Cookies', () => {
@@ -23,34 +26,36 @@ describe('Measure Packager: Validate before packaging a FHIR measure', () => {
 
     it('QDM Measure: Validate error message on measure packager page', () => {
 
+        helper.enabledWithTimeout(measurelibrary.searchInputBox)
         helper.enterText(measurelibrary.searchInputBox, qdmMeasure)
-        cy.get(measurelibrary.searchBtn).click();
+        cy.get(measurelibrary.searchBtn).click()
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        cy.get(measurelibrary.row1MeasureSearch).dblclick();
+        cy.get(measurelibrary.row1MeasureSearch).dblclick()
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        cy.get(measureComposer.cqlWorkspace).click();
+        cy.get(measureComposer.cqlWorkspace).click()
 
         helper.verifySpinnerAppearsAndDissappears()
 
         // Function
-        cy.get(measureComposer.functionMeasureComposer).click();
+        cy.get(measureComposer.functionMeasureComposer).click()
         helper.waitToContainText(measureComposer.cqlWorkspaceTitleGlobal2, 'Function')
-        cy.get(measureComposer.functionNameInput).type('TestForValidationError', { delay: 50 });
-        cy.get(measureComposer.functionCQLExpressionEditorInput).type('dfgfkj', { delay: 50 });
-        cy.get(measureComposer.functionSaveBtn).click();
-
+        cy.get(measureComposer.functionNameInput).type('TestForValidationError', { delay: 50 })
+        cy.get(measureComposer.functionCQLExpressionEditorInput).type('dfgfkj', { delay: 50 })
+        cy.pause()
+        cy.get(measureComposer.functionSaveBtn).click()
+cy.pause()
         helper.verifySpinnerAppearsAndDissappears()
 
         // Measure Packager 
-        cy.get(measureComposer.measurePackager).click();
+        cy.get(measureComposer.measurePackager).click()
         helper.verifySpinnerAppearsAndDissappears()
         cy.get(measureComposer.packageWarningMessage).should('contain.text', 'Your CQL file contains validation errors. Errors must be corrected before proceeding to measure packaging. Please return to the CQL Workspace to make corrections.');
 
-        cy.get(measurelibrary.measureLibraryTab).click();
+        cy.get(measurelibrary.measureLibraryTab).click()
 
         helper.verifySpinnerAppearsAndDissappears()
