+++ cypress/integration/MAT/WebInterface/MeasureComposer/MeasurePackager/PackagingContinuousVariableFhirMeasure.spec.js
@@ -1,4 +1,5 @@
-import * as helper from '../../../../../support/helpers';
+import * as helper from '../../../../../support/helpers'
+import * as dataCreation from '../../../../../support/MAT/MeasureAndCQLLibraryCreation'
 import * as measurelibrary from '../../../../../pom/MAT/WI/MeasureLibrary'
 import * as oktaLogin from '../../../../../support/oktaLogin'
 import * as createNewMeasure from '../../../../../pom/MAT/WI/CreateNewMeasure'
@@ -66,9 +67,9 @@ describe('Packaging: Continuous Variable Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addValueSet('2.16.840.1.113883.3.666.5.307')
-        helper.addValueSet('2.16.840.1.113762.1.4.1182.118')
-        helper.addValueSet('2.16.840.1.113762.1.4.1111.161')
+        dataCreation.addValueSet('2.16.840.1.113883.3.666.5.307')
+        dataCreation.addValueSet('2.16.840.1.113762.1.4.1182.118')
+        dataCreation.addValueSet('2.16.840.1.113762.1.4.1111.161')
 
         // Codes
 
@@ -76,9 +77,9 @@ describe('Packaging: Continuous Variable Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addCode('CODE:/CodeSystem/LOINC/Version/2.46/Code/21112-8/Info')
-        helper.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2016-03/Code/419099009/Info')
-        helper.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2017-09/Code/371828006/Info')
+        dataCreation.addCode('CODE:/CodeSystem/LOINC/Version/2.46/Code/21112-8/Info')
+        dataCreation.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2016-03/Code/419099009/Info')
+        dataCreation.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2017-09/Code/371828006/Info')
 
         // Definition
 
@@ -86,8 +87,8 @@ describe('Packaging: Continuous Variable Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addDefinition('Initial Population', 'TJC."Encounter with Principal Diagnosis and Age"')
-        helper.addDefinition('Measure Population', '"Initial Population"')
+        dataCreation.addDefinition('Initial Population', 'TJC."Encounter with Principal Diagnosis and Age"')
+        dataCreation.addDefinition('Measure Population', '"Initial Population"')
 
         // Function
