+++ cypress/integration/MAT/WebInterface/MeasureComposer/MeasurePackager/PackagingProportionFhirMeasure.spec.js
@@ -1,4 +1,5 @@
-import * as helper from '../../../../../support/helpers';
+import * as helper from '../../../../../support/helpers'
+import * as dataCreation from '../../../../../support/MAT/MeasureAndCQLLibraryCreation'
 import * as measurelibrary from '../../../../../pom/MAT/WI/MeasureLibrary'
 import * as oktaLogin from '../../../../../support/oktaLogin'
 import * as createNewMeasure from '../../../../../pom/MAT/WI/CreateNewMeasure'
@@ -19,6 +20,9 @@ describe('Packaging: Proportion Measure', () => {
 
     it('Validate the measure packaging for Proportion FHIR Measure', () => {
 
+        helper.verifySpinnerAppearsAndDissappears()
+
+        helper.enabledWithTimeout(measurelibrary.newMeasureButton)
         cy.get(measurelibrary.newMeasureButton).click()
         let measureName = 'CreateFhirProportionMeasure' + Date.now()
 
@@ -65,9 +69,9 @@ describe('Packaging: Proportion Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addValueSet('2.16.840.1.113883.3.666.5.307')
-        helper.addValueSet('2.16.840.1.113762.1.4.1182.118')
-        helper.addValueSet('2.16.840.1.113762.1.4.1111.161')
+        dataCreation.addValueSet('2.16.840.1.113883.3.666.5.307')
+        dataCreation.addValueSet('2.16.840.1.113762.1.4.1182.118')
+        dataCreation.addValueSet('2.16.840.1.113762.1.4.1111.161')
 
         // Codes
 
@@ -75,9 +79,9 @@ describe('Packaging: Proportion Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2020-03/Code/420523002/Info')
-        helper.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2016-03/Code/419099009/Info')
-        helper.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2017-09/Code/371828006/Info')
+        dataCreation.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2020-03/Code/420523002/Info')
+        dataCreation.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2016-03/Code/419099009/Info')
+        dataCreation.addCode('CODE:/CodeSystem/SNOMEDCT/Version/2017-09/Code/371828006/Info')
 
         // Definition
 
@@ -85,9 +89,9 @@ describe('Packaging: Proportion Measure', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        helper.addDefinition('Initial Population', 'TJC."Encounter with Principal Diagnosis and Age"')
-        helper.addDefinition('Denominator', 'TJC."Ischemic Stroke Encounter"')
-        helper.addDefinition('Numerator', 'true')
+        dataCreation.addDefinition('Initial Population', 'TJC."Encounter with Principal Diagnosis and Age"')
+        dataCreation.addDefinition('Denominator', 'TJC."Ischemic Stroke Encounter"')
+        dataCreation.addDefinition('Numerator', 'true')
 
         //CQL Library Editor
