+++ cypress/integration/MAT/WebInterface/MeasureLibrary/Grid.spec.js
@@ -1,4 +1,5 @@
 import * as helper from '../../../../support/helpers'
+import * as dataCreation from '../../../../support/MAT/MeasureAndCQLLibraryCreation'
 import * as measurelibrary from '../../../../pom/MAT/WI/MeasureLibrary'
 import * as createNewMeasure from "../../../../pom/MAT/WI/CreateNewMeasure";
 import * as measureComposer from "../../../../pom/MAT/WI/MeasureComposer";
@@ -17,10 +18,10 @@ describe('Measure Library Row Selection', () => {
         oktaLogin.login()
 
         //creating new draft measure
-        draftMeasure = helper.createDraftMeasure()
+        draftMeasure = dataCreation.createDraftMeasure()
 
         //creating new versioned measure
-        versionMeasure = helper.createMajorVersionMeasure()
+        versionMeasure = dataCreation.createMajorVersionMeasure()
 
         cy.get(measurelibrary.row1MeasureSearch).click()
 
@@ -122,17 +123,17 @@ describe('Measure Library Row Selection', () => {
 
 describe('Measure Library Recent Activity Grid', () => {
     before('Login', () => {
-        versionMeasureNotOwner = helper.loginCreateVersionedMeasureNotOwnerLogout()
+        versionMeasureNotOwner = dataCreation.loginCreateVersionedMeasureNotOwnerLogout()
 
         oktaLogin.login()
 
         //creating new draft measure
-        draftMeasure = helper.createDraftMeasure()
+        draftMeasure = dataCreation.createDraftMeasure()
 
         //creating new versioned measure
-        versionMeasure = helper.createMajorVersionMeasure()
+        versionMeasure = dataCreation.createMajorVersionMeasure()
 
-        fhirMeasure = helper.createDraftMeasure('FhirDraft','FHIR')
+        fhirMeasure = dataCreation.createDraftMeasure('FhirDraft','FHIR')
     })
     beforeEach('Preserve Cookies', () => {
         helper.preserveCookies()
@@ -185,7 +186,7 @@ describe('Measure Library Recent Activity Grid', () => {
 
     it('Enabled/Disabled Recent Activity The Owner', () => {
 
-        measureName = helper.createDraftMeasure('TestMeasure')
+        measureName = dataCreation.createDraftMeasure('TestMeasure')
 
         helper.enterText(measurelibrary.searchInputBox, measureName)
 
@@ -459,23 +460,25 @@ describe('Measure Library Recent Activity Grid', () => {
 describe('Measure Library Grid Button Bar', () => {
     before('Login', () => {
 
-        versionMeasureNotOwner = helper.loginCreateVersionedMeasureNotOwnerLogout()
+        versionMeasureNotOwner = dataCreation.loginCreateVersionedMeasureNotOwnerLogout()
 
         oktaLogin.login()
 
         helper.verifySpinnerAppearsAndDissappears()
 
         //creating new draft measure
-        draftMeasure = helper.createDraftMeasure()
+        draftMeasure = dataCreation.createDraftMeasure()
 
         helper.verifySpinnerAppearsAndDissappears()
 
         //creating new versioned measure
-        versionMeasure = helper.createMajorVersionMeasure()
+        versionMeasure = dataCreation.createMajorVersionMeasure()
 
         helper.verifySpinnerAppearsAndDissappears()
 
-        fhirMeasure = helper.createDraftMeasure('FhirDraft','FHIR')
+        fhirMeasure = dataCreation.createDraftMeasure('FhirDraft','FHIR')
+
+        helper.verifySpinnerAppearsAndDissappears()
 
     })
     beforeEach('Preserve Cookies', () => {
@@ -496,7 +499,9 @@ describe('Measure Library Grid Button Bar', () => {
         cy.get(measurelibrary.searchBtn).click()
 
         helper.verifySpinnerAppearsAndDissappears()
+        helper.verifySpinnerAppearsAndDissappears()
 
+        helper.visibleWithTimeout(measurelibrary.row1MeasureSearch)
         cy.get(measurelibrary.row1MeasureSearch).click()
 
         helper.disabled(measurelibrary.createVersionDraftMeasureSearchBtn)
@@ -513,15 +518,18 @@ describe('Measure Library Grid Button Bar', () => {
 
     it('Enabled/Disabled Measure Search Table The Owner', () => {
 
-        measureName = helper.createDraftMeasure('TestMeasure')
+        measureName = dataCreation.createDraftMeasure('TestMeasure')
 
+        helper.verifySpinnerAppearsAndDissappears()
+
+        helper.enabledWithTimeout(measurelibrary.searchInputBox)
         helper.enterText(measurelibrary.searchInputBox, measureName)
         cy.get(measurelibrary.searchBtn).click()
 
         helper.verifySpinnerAppearsAndDissappears()
+        helper.verifySpinnerAppearsAndDissappears()
 
-        cy.wait(2000)
-
+        helper.visibleWithTimeout(measurelibrary.row1MeasureSearch)
         cy.get(measurelibrary.row1MeasureSearch).click()
 
         helper.enabled(measurelibrary.createVersionMeasureSearchBtn)
@@ -539,9 +547,9 @@ describe('Measure Library Grid Button Bar', () => {
         cy.get(measurelibrary.continueBtn).click()
 
         helper.verifySpinnerAppearsAndDissappears()
+        helper.verifySpinnerAppearsAndDissappears()
 
-        cy.wait(2000)
-
+        helper.visibleWithTimeout(measurelibrary.row1MeasureSearch)
         cy.get(measurelibrary.row1MeasureSearch).click()
 
         helper.enabled(measurelibrary.createDraftMeasureSearchBtn)
@@ -558,9 +566,9 @@ describe('Measure Library Grid Button Bar', () => {
         cy.get(measurelibrary.searchBtn).click()
 
         helper.verifySpinnerAppearsAndDissappears()
+        helper.verifySpinnerAppearsAndDissappears()
 
-        cy.wait(2000)
-
+        helper.visibleWithTimeout(measurelibrary.row1MeasureSearch)
         cy.get(measurelibrary.row1MeasureSearch).click()
 
         helper.enabled(measurelibrary.createVersionMeasureSearchBtn)
@@ -580,6 +588,7 @@ describe('Measure Library Grid Button Bar', () => {
 
         helper.verifySpinnerAppearsAndDissappears()
 
+        helper.visibleWithTimeout(measurelibrary.row1MeasureSearch)
         cy.get(measurelibrary.row1MeasureSearch).click()
 
         cy.get(measurelibrary.createVersionMeasureSearchBtn).click()
