+++ cypress/support/index.js
@@ -17,6 +17,9 @@
 import './commands'
 import 'cypress-axe'
 
+// Retry failed tests:
+require('cypress-plugin-retries')
+
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
