+++ package.json
@@ -24,7 +24,7 @@
         "compile:utils": "tsc -p ./utils/tsconfig.json;",
         "lint:tslint": "tslint -c tslint.json './cypress/integration/**/*spec.ts' 'utils/**/*.ts'",
         "dev:all:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/**/*.spec.js' --browser chrome;",
-        "TEST": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=matDev --spec 'cypress/integration/MAT/WebInterface/CQL_Library/FHIRLibraryConversion/*.spec.js' --browser chrome;",
+        "TEST": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=matDev --spec 'cypress/integration/MAT/WebInterface/CQL_Library/VersioningStandaloneLibrary.spec.js' --browser chrome;",
         "matdev:all:ui:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=matDev --spec 'cypress/integration/MAT/WebInterface/**/*.spec.js' --browser chrome;",
         "dev:all:ui:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/WebInterface/**/*.spec.js' --browser chrome;",
         "all:service:tests": "NO_COLOR=1 $(npm bin)/cypress run --env configFile=dev --spec 'cypress/integration/MAT/Services/**/*.spec.js' --browser chrome;",
@@ -55,6 +55,7 @@
         "axe-core": "^3.4.0",
         "cypress": "^4.5.0",
         "cypress-axe": "^0.5.1",
+        "cypress-plugin-retries": "^1.5.2",
         "data-driven": "^1.4.0",
         "eslint-plugin-chai-friendly": "^0.5.0",
         "eslint-plugin-cypress": "^2.7.0",
@@ -68,7 +69,7 @@
     "dependencies": {
         "@cypress/webpack-preprocessor": "^4.1.1",
         "@types/node": "^12.12.14",
-        "cypress-visual-regression": "^1.0.5",
+        "cypress-visual-regression": "^1.3.1",
         "mocha": "5.2.0",
         "mochawesome": "^4.1.0",
         "mochawesome-merge": "^4.0.0",
